# Optimize project deletion performance with batch operations and incremental stats